+++ features/productHub/helpers/getActionUrl.ts
@@ -47,7 +47,7 @@ export function getActionUrl({
         ? ProductHubProductType.Multiply
         : product
 
-      return `/ajna/${productInUrl}/${label.replace('/', '-')}`
+      return `/ethereum/ajna/${productInUrl}/${label.replace('/', '-')}`
     case LendingProtocol.AaveV2:
       return getAaveStrategyUrl({
         aaveVersion: 'v2',
