+++ pages/ajna/position/[id].tsx
@@ -1,27 +1,28 @@
-import { AjnaLayout, ajnaPageSeoTags } from 'features/ajna/common/layout'
-import { AjnaProductController } from 'features/ajna/positions/common/controls/AjnaProductController'
 import { GetServerSidePropsContext } from 'next'
-import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import React from 'react'
 
-interface AjnaManagePositionPageProps {
-  id: string
+function AjnaManagePositionPageRedirect() {
+  return <></>
 }
 
-function AjnaManagePositionPage({ id }: AjnaManagePositionPageProps) {
-  return <AjnaProductController id={id} flow="manage" />
-}
+export default AjnaManagePositionPageRedirect
 
-AjnaManagePositionPage.layout = AjnaLayout
-AjnaManagePositionPage.seoTags = ajnaPageSeoTags
+export async function getServerSideProps({ query }: GetServerSidePropsContext) {
+  const id = query.id as string
 
-export default AjnaManagePositionPage
+  if (!isNaN(parseInt(id, 10))) {
+    return {
+      redirect: {
+        destination: `/ethereum/ajna/${id}`,
+        permanent: true,
+      },
+    }
+  }
 
-export async function getServerSideProps({ locale, query }: GetServerSidePropsContext) {
   return {
-    props: {
-      ...(await serverSideTranslations(locale || 'en', ['common'])),
-      id: query.id,
+    redirect: {
+      permanent: false,
+      destination: '/not-found',
     },
   }
 }
